+++ IPython/nbconvert/filters/highlight.py
@@ -109,7 +109,7 @@ def _pygments_highlight(source, output_formatter, language='ipython', metadata=N
     from pygments import highlight
     from pygments.lexers import get_lexer_by_name
     from pygments.util import ClassNotFound
-    from IPython.nbconvert.utils.lexers import IPythonLexer, IPython3Lexer
+    from IPython.lib.lexers import IPythonLexer, IPython3Lexer
 
     # If the cell uses a magic extension language,
     # use the magic language instead.
